medical_fusion_webapp/requirements.txt
Flask==2.3.3
torch>=1.9.0
torchvision>=0.10.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-image==0.21.0
pytorch-wavelets==1.3.0
pandas==2.0.3
Werkzeug==2.3.7